applyimage
This python script duplicates an image in photoshop and applies a subtract filter to the image under the apply image dialogue box. It then applies a linear light blending mode to the layer in the layer blending dialogue box. This script should be used in conjunction with the gaussian blur script I wrote to automate the retouching process in Adobe Photoshop.